SCB-172 fix tests for upgrading to zipkin2
The name, traceId, id members are refactored to methods in zipkin2. And the
binaryAnnotations member is refactored to method tags() which returns a map.
The dependency on io.zipkin.java artivacts are not needed anymore. But the
mocking in tests requires zipkin.Endpoint so we kept it for test scope.